requirements.txt
absl-py==2.3.0
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
bitsandbytes==0.46.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
deepspeed==0.16.9
dill==0.3.8
docstring_parser==0.16
einops==0.8.1
evaluate==0.4.4
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2024.9.0
gradio==5.12.0
gradio_client==1.5.4
h11==0.16.0
hf_transfer==0.1.9
hf-xet==1.1.2
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
ijson==3.4.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llamafactory==0.9.2
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.6
mdurl==0.1.2
modelscope==1.28.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pfzy==0.3.4
pillow==11.2.1
pip==25.1
platformdirs==4.3.7
pooch==1.8.2
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==6.31.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rouge-chinese==1.0.3
rouge_score==0.1.2
ruff==0.11.12
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==78.1.1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.3.6
stack-data==0.2.0
starlette==0.46.2
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.3.0
trl==0.9.6
typer==0.16.0
typing_extensions==4.12.2
typing-inspection==0.4.1
tyro==0.8.14
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==14.2
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
